Extract replace text function to new file
