+++ app/javascript/app/components/abbr-replace/abbr-replace.js
@@ -1,4 +1,4 @@
 import Component from './abbr-replace-component';
 
-export { replaceStringAbbr } from './abbr-replace-component';
+export { replaceStringAbbr } from './replace-string-abbr';
 export default Component;
